+++ src/Frontend/Core/Engine/Page.php
@@ -107,12 +107,6 @@ class Page extends KernelLoader
      */
     public function load(): void
     {
-        // @deprecated remove this in Fork 6, the privacy consent dialog should be used
-        if (!$this->getContainer()->get('fork.settings')->get('Core', 'show_consent_dialog', false)) {
-            // set tracking cookie
-            Model::getVisitorId();
-        }
-
         // create header instance
         $this->header = new Header($this->getKernel());
 
@@ -193,13 +187,6 @@ class Page extends KernelLoader
             $this->template->assignGlobal('isPage' . $this->pageId, true);
             $this->template->assignGlobal('isChildOfPage' . $this->record['parent_id'], true);
 
-            // hide the cookiebar from within the code to prevent flickering
-            // @deprecated remove this in Fork 6, the privacy consent dialog should be used
-            $this->template->assignGlobal(
-                'cookieBarHide',
-                !$this->get('fork.settings')->get('Core', 'show_cookie_bar', false)
-                || $this->getContainer()->get('fork.cookie')->hasHiddenCookieBar()
-            );
             $this->parsePrivacyConsents();
             $this->parsePositions();
 
@@ -340,9 +327,19 @@ class Page extends KernelLoader
     {
         $consentDialog = $this->get(ConsentDialog::class);
 
+        $levels = array_map(
+            function (string $level): array {
+                return [
+                    'name' => $level,
+                    'localeKey' => str_replace(' ', '', ucwords(str_replace('_', ' ', $level))),
+                ];
+            },
+            $consentDialog->getLevels()
+        );
+
         $this->template->assignGlobal('privacyConsentEnabled', $consentDialog->isDialogEnabled());
-        $this->template->assignGlobal('privacyConsentDialogHide', !$consentDialog->shouldDialogBeShown());
-        $this->template->assignGlobal('privacyConsentDialogLevels', $consentDialog->getLevels());
+        $this->template->assignGlobal('privacyConsentDialogShow', $consentDialog->shouldDialogBeShown());
+        $this->template->assignGlobal('privacyConsentDialogLevels', $levels);
     }
 
     protected function parsePositions(): void